pyproject.toml
[project]
name = "dodo-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.1",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
